refactor: remove dependency on AnotherWorld (#38)
GrowingFlora currently depends on AnotherWorld, but there isn't much reason for that to be the case. I moved some utilities here and to the engine (see MovingBlocks/Terasology#4849, which this PR depends on) and updated some references, which was enough to remove the dependency. Behavior shouldn't change, except that it no longer considers parent biomes for AnotherWorldsBiomes, and it also no longer has a bug where branch angles were snapped to whole numbers of radians.

This also makes the dependency on ClimateConditions optional so that climate still affects growth if it is present.

* chore: remove dependency on AnotherWorld
* fix: add border to tree facet
* chore: distribution utilities were moved
* refactor: make ClimateConditions dependency optional
* chore: increase use of Vector3ic