fix: planted saplings will actually grow (#40)
Planted saplings may have grown in the past a small percentage of the time, but now they do it reliably. Try testing with, e.g. with MetalRenegades, `give BaobabSapling` or `give MetalRenegades:CypressSapling`, plant it, probably increase the time dilation, and wait a bit.

Trees that generate in the world almost always still don't grow, and that's partially because `OnActivatedComponent` events aren't sent until something interacts with a block with an attached entity (e.g. looking at it, this is what causes #2) and partially because of other problems I haven't figured out yet.